add index files tj page results and rename style files
